maryco-backend/requirements.txt
arrow==1.4.0
asgiref==3.11.1
attrs==26.1.0
binaryornot==0.6.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cookiecutter==2.7.1
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-resized==1.0.3
django-tailwind==4.4.2
djangorestframework==3.17.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pillow==12.1.1
psycopg2-binary==2.9.11
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
